Feature: Support 64-bit terrain masks

+++ apps/src/test/scala/com/crib/bills/dom6maps/Arbitraries.scala
@@ -90,7 +90,7 @@ object Arbitraries:
   given Arbitrary[Terrain] =
     Arbitrary(for
       p <- summon[Arbitrary[ProvinceId]].arbitrary
-      m <- Gen.choose(0, Int.MaxValue)
+      m <- Gen.choose(0L, Long.MaxValue)
     yield Terrain(p, m))
 
   given Arbitrary[LandName] =
